requirements.txt
# Минимальный набор зависимостей для локального RAG пайплайна

chromadb>=0.5.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.36.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Дополнительные инструменты можно установить при необходимости:
# pip install ".[dev]" или вручную (pytest, black, jupyter и т.п.)
